Made the supported Markup tags explicit. Formatted inline code.
This resolves an issue with combinations of brackets and paranthesis
which normally would format a link in Markdown but have no effect in
Discord being partially erased. This incidentally removes list formatting
and a number of other markdown features not supported by Discord.

Inline code is forcibly formatted as white-on-black and given a single
grave wrapper. This isn't pretty in the end, but what was there wasn't
much better.

Lastly, blockquotes are not formatted as of this patch.